backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.7
flask-limiter==3.7.0
flask-caching==2.1.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.5
openai==1.10.0
redis==5.0.1
structlog==24.1.0
bleach==6.1.0
pillow==10.0.0
gunicorn==22.0.0
gevent==24.2.1